+++ rector.php
@@ -2,12 +2,15 @@
 declare(strict_types=1);
 
 use craft\rector\SetList as CraftSetList;
+
 use Rector\Set\ValueObject\LevelSetList;
 use Rector\Set\ValueObject\SetList;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Rector\Config\RectorConfig;
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $containerConfigurator->import(CraftSetList::CRAFT_CMS_40);
-    $containerConfigurator->import(SetList::CODE_QUALITY);
-    $containerConfigurator->import(LevelSetList::UP_TO_PHP_81);
+return static function (RectorConfig $rectorConfig): void {
+    $rectorConfig->sets([
+        CraftSetList::CRAFT_CMS_40,
+        SetList::CODE_QUALITY,
+        LevelSetList::UP_TO_PHP_81
+    ]);
 };