+++ src/helpers/ApiHelper.php
@@ -13,7 +13,7 @@ class ApiHelper
 {
 
     /** @var \string[][] */
-    const LANGUAGE_CODE_MAP = [
+    final const LANGUAGE_CODE_MAP = [
         'nb' => ['nb-NO', 'nn', 'nn-NO', 'no'],
     ];