+++ src/models/Settings.php
@@ -8,17 +8,26 @@ class Settings extends Model
 {
 
     /** @var string|null */
-    public ?string $damUrl;
+    public ?string $damUrl = null;
 
     /** @var string|null */
-    public ?string $jwtSecret;
+    public ?string $jwtSecret = null;
 
-    /** @var array|null */
+    /** @var string|null */
     public ?array $metaDataFieldMap;
 
-    /** @var float */
-    public float $hlsVideoLazyloadDelay = 0;
+    /** @var float|int */
+    public float|int $hlsVideoLazyloadDelay = 0;
 
     /** @var string */
     public string $hlsVideoLazyloadEvent = 'DOMContentLoaded';
+
+    /** @var string|null */
+    public ?string $pluginName = null;
+
+    /** @var string|null */
+    public ?string $cpSectionPath = null;
+
+    /** @var string[]|null */
+    public ?array $allowedVolumesForImport = null;
 }