+++ src/records/ImportedFile.php
@@ -23,8 +23,6 @@ class ImportedFile extends ActiveRecord
 {
     /**
      * @inheritdoc
-     *
-     * @return string
      */
     public static function tableName(): string
     {